APP - mailing

+++ API/cron.php
@@ -12,13 +12,15 @@
         $count   = 10;
         $inArray = '';
         $emails  = $db->query("SELECT * FROM `mailing_cron` LIMIT 0,".$count, []);
-        for ($i=0; $i<$count; $i++){
-            $inArray .= "'{$emails[$i]['email']}', ";
-            mail($emails[$i]['email'], $emails[$i]['theme'], $emails[$i]['content']);
+        if ($emails){
+            for ($i=0; $i<$count; $i++){
+                $inArray .= "'{$emails[$i]['email']}', ";
+                mail($emails[$i]['email'], $emails[$i]['theme'], $emails[$i]['content']);
+            }
+            $inArray = substr($inArray, 0, -2);
+            $inArray = '('.$inArray.')';
+            $db->query("DELETE FROM `mailing_cron` WHERE `email` IN ".$inArray, []);
         }
-        $inArray = substr($inArray, 0, -2);
-        $inArray = '('.$inArray.')';
-        $db->query("DELETE FROM `mailing_cron` WHERE `email` IN ".$inArray, []);
         $db->disconnect();
     }
 ?>
